copy_output(pycenterpoint ./) tensorrt
